+++ src/xyz/oihana/schema/places/ProviderSite.php
@@ -2,6 +2,7 @@
 
 namespace xyz\oihana\schema\places;
 
+use xyz\oihana\schema\traits\places\SiteFlagsTrait;
 use xyz\oihana\schema\traits\SiteTrait;
 
 /**
@@ -21,5 +22,6 @@ use xyz\oihana\schema\traits\SiteTrait;
  */
 class ProviderSite extends Site
 {
-    use SiteTrait ;
+    use SiteFlagsTrait ,
+        SiteTrait      ;
 }